refactor: consistent named exports for layout components
