+++ src/app/[locale]/newsroom/[slug]/page.tsx
@@ -8,7 +8,7 @@ import { getPostBySlug, listPostSlugs } from '@/lib/posts';
 import { sitePageMetadata } from '@/lib/site-page-metadata';
 
 import { Container } from '@/components/layout/Container';
-import Page from '@/components/layout/Page';
+import { Page } from '@/components/layout';
 import { NewsArticle } from '@/components/templates/NewsArticle';
 import { ArrowLink } from '@/components/ui';
